Revert "gh:552 - Corrected Series.mean return type (#622)"
This reverts commit c03e23c1b84e3f568e1f2887d1c49b653b023bf3.

+++ pandas-stubs/core/series.pyi
@@ -1609,7 +1609,7 @@ class Series(IndexOpsMixin, NDFrame, Generic[S1]):
         level: None = ...,
         numeric_only: _bool = ...,
         **kwargs,
-    ) -> np.float64: ...
+    ) -> float: ...
     def median(
         self,
         axis: AxisIndex | None = ...,
